refactor products' form
